+++ sub_locacoes/lib/engine/CRUD-Fire.dart
@@ -1,9 +1,49 @@
 import 'package:flutter/material.dart';
 import 'package:sub_locacoes/engine/xrud.dart';
+import 'package:sub_locacoes/engine/read-demo.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
 
 void main() {
   runApp(MyApp()); //Enviando commit
+}
+
+class ReadDocument extends StatelessWidget {
+  final String documentId;
+  final String collection;
+
+  ReadDocument(this.documentId, this.collection);
+
+  @override
+  Widget build(BuildContext context) {
+    CollectionReference users =
+        FirebaseFirestore.instance.collection(collection);
+
+    return FutureBuilder<DocumentSnapshot>(
+      future: users.doc(documentId).get(),
+      builder:
+          (BuildContext context, AsyncSnapshot<DocumentSnapshot> snapshot) {
+        if (snapshot.hasError) {
+          String mensageiro = "Um erro ocorreu durante a leitura do sistema";
+          return ListBody(
+            children: <Widget>[
+              Text(mensageiro),
+              // Text('Would you like to approve of this message?'),
+            ],
+          );
+        }
+
+        if (snapshot.connectionState == ConnectionState.done) {
+          Map<String, dynamic> retorno = snapshot.data.data();
+          return Text(
+              "Olha a novidade aí gente: Full Name: ${retorno['full_name']} Data: ${retorno['age']}");
+        }
+
+        return Text("loading");
+      },
+    );
+  }
 }
 
 final FirebaseAuth _auth = FirebaseAuth.instance;
@@ -102,6 +142,7 @@ class _MyCRUDPageState extends State<MyCRUDPage> {
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
+            new ReadDocument("finalmente", "users"),
             Text(
               'You have pushed the button this many times:',
             ),
